service/requirements.txt
Django>=3.2.9,<4.0
django-cors-headers
requests
psycopg2-binary>=2.8
djangorestframework==3.13.1
